+++ src/_models/Role.ts
@@ -1,6 +1,9 @@
-export type RoleType = 'admin' | 'front';
+import { EmailAddress } from './User';
+
+export const nameOfRole = ['admin', 'front'] as const;
+export type RoleType = typeof nameOfRole[number];
 export type Role = {
   id: string;
   name: RoleType;
-  members: string[];
+  members: EmailAddress[];
 };