+++ src/master/role/DataInputs.tsx
@@ -2,6 +2,8 @@ import { useTranslation } from 'react-i18next';
 import { Control, DeepMap, DeepPartial, FieldError, FieldValues, Path } from 'react-hook-form';
 
 import { ControllerTextField } from '_components/ControllerTextField';
+import { AddrBookAutoComplete } from '_components/AddrBookAutoComplete';
+import { nameOfRole } from '_models/Role';
 
 type DataInputsProps<TFieldValues extends FieldValues> = {
   control: Control<TFieldValues>;
@@ -15,15 +17,17 @@ export function DataInputs<TFieldValues extends FieldValues>(props: DataInputsPr
 
   return (
     <>
-      <ControllerTextField name={'name' as Path<TFieldValues>} control={control} label={t('settings.header.role.name')} required errors={errors} />
       <ControllerTextField
-        name={'members' as Path<TFieldValues>}
+        name={'name' as Path<TFieldValues>}
         control={control}
-        label={t('settings.header.role.members')}
+        label={t('settings.header.role.name')}
         required
-        multiline
+        selectList={nameOfRole.map((value) => {
+          return { label: t(`settings.view.role.name.${value}`), value: value };
+        })}
         errors={errors}
       />
+      <AddrBookAutoComplete name={'members' as Path<TFieldValues>} control={control} label={t('settings.header.role.members')} errors={errors} />
     </>
   );
 }